+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed indentation of comments around statements outside of statement lists.
+- Fixed performance issues with heavily nested conditional directives.
 
 ## [0.5.1] - 2025-06-02